Refactor usePokemon  and create services folder to hold fetching logic
