+++ src/hooks/usePokemon.ts
@@ -1,14 +1,6 @@
 import { useEffect, useState } from "react";
 import type { Pokemon } from "../types/types";
-import { getUrlId } from "../utils/getUrlId";
-
-interface RawPokemon {
-  sprites: { front_default: string };
-  types: { type: { name: string } }[];
-  weight: number;
-  height: number;
-  stats: { base_stat: number; stat: { name: string } }[];
-}
+import { pokemonService } from "../services/pokemonService";
 
 export function usePokemon() {
   const [pokemons, setPokemons] = useState<Pokemon[]>([]);
@@ -21,35 +13,7 @@ export function usePokemon() {
         setLoading(true);
         setError("");
 
-        const response = await fetch("https://pokeapi.co/api/v2/pokemon?limit=50");
-        if (!response.ok) throw new Error(`Error fetching pokémon! Try again`);
-        const data = await response.json();
-
-        const pokemons = await Promise.all(
-          data.results.map(async (pokemon: { url: string, name: string }) => {
-            const response = await fetch(pokemon.url);
-            if (!response.ok) throw new Error(`Error fetching ${pokemon.name}`);
-            const d: RawPokemon = await response.json();
-            const uId = getUrlId(pokemon.url)
-
-            return {
-              id: parseInt(uId),
-              name: pokemon.name,
-              image: d.sprites.front_default,
-              types: d.types.map((t: any) => ({
-                name: t.type.name
-              })),
-              weight: d.weight,
-              height: d.height,
-              stats: d.stats.filter((s: any) => ["hp", "attack", "defense", "speed"].includes(s.stat.name))
-                .map((s: any) => ({
-                  name: s.stat.name,
-                  base: s.base_stat,
-                })),
-            };
-          })
-        );
-
+        const pokemons = await pokemonService.fetchPokemonList(50);
         setPokemons(pokemons);
       } catch (err) {
         if (err instanceof Error) {
